+++ @xen-orchestra/backups/writers/_checkVhd.js
@@ -1,4 +1,4 @@
-const Vhd = require('vhd-lib').default
+const Vhd = require('vhd-lib').VhdFile
 
 exports.checkVhd = async function checkVhd(handler, path) {
   await new Vhd(handler, path).readHeaderAndFooter()