+++ packages/xo-web/src/xo-app/xoa/licenses/index.js
@@ -139,7 +139,7 @@ const PRODUCTS_COLUMNS = [
   },
   {
     name: _('licenseExpires'),
-    itemRenderer: ({ expires }) => (expires !== undefined ? <ShortDate timestamp={expires} /> : '-'),
+    itemRenderer: ({ expires }) => (expires !== undefined ? <ShortDate timestamp={+expires} /> : '-'),
     sortCriteria: 'expires',
     sortOrder: 'desc',
   },